+++ outputs.tf
@@ -124,3 +124,34 @@ output "link_full_object" {
   value       = var.link.create ? awscc_rtbfabric_link.link[0] : null
 }
 
+
+# Inbound External Link Outputs
+output "inbound_external_link_id" {
+  description = "Link ID of the created inbound external link"
+  value       = var.inbound_external_link.create ? awscc_rtbfabric_inbound_external_link.inbound_external_link[0].link_id : null
+}
+
+output "inbound_external_link_arn" {
+  description = "ARN of the created inbound external link"
+  value       = var.inbound_external_link.create ? awscc_rtbfabric_inbound_external_link.inbound_external_link[0].arn : null
+}
+
+output "inbound_external_link_status" {
+  description = "Status of the created inbound external link"
+  value       = var.inbound_external_link.create ? awscc_rtbfabric_inbound_external_link.inbound_external_link[0].link_status : null
+}
+
+output "inbound_external_link_gateway_id" {
+  description = "Gateway ID that the inbound external link is attached to"
+  value       = var.inbound_external_link.create ? awscc_rtbfabric_inbound_external_link.inbound_external_link[0].gateway_id : null
+}
+
+output "inbound_external_link_created_timestamp" {
+  description = "Creation timestamp of the inbound external link"
+  value       = var.inbound_external_link.create ? awscc_rtbfabric_inbound_external_link.inbound_external_link[0].created_timestamp : null
+}
+
+output "inbound_external_link_updated_timestamp" {
+  description = "Last update timestamp of the inbound external link"
+  value       = var.inbound_external_link.create ? awscc_rtbfabric_inbound_external_link.inbound_external_link[0].updated_timestamp : null
+}